Add initial deployment code using Heat
Add code to deploy a Heat stack for the bastion host. No configuration
is done inside the virtual machine yet - it is simply deployed in
OpenStack. The code can be used to generate one or more randomly named
bastion hosts inside an OpenStack project and then to update them by
referring to them by name. The name for the stack can also be given by
the user.

Also add README.md file describing purpose and usage.